streamline WS server logging
