+++ lma-websocket-stack/source/app/src/calleventdata/eventtypes.ts
@@ -7,7 +7,6 @@ import {
     CategoryEvent,
 } from '@aws-sdk/client-transcribe-streaming';
 import stream from 'stream';
-import { CallMetaData } from './lca';
 import { WriteStream } from 'fs';
 
 export type Uuid = string;             // UUID as defined by RFC#4122
@@ -58,6 +57,17 @@ export type AddTranscriptSegmentEvent = CallEventBase<'ADD_TRANSCRIPT_SEGMENT'> 
 
 export type AddCallCategoryEvent = CallEventBase<'ADD_CALL_CATEGORY'> & {
     CategoryEvent: CategoryEvent,
+};
+
+export type CallMetaData = {
+    callId: Uuid,
+    fromNumber?: string,
+    toNumber?: string,
+    shouldRecordCall?: boolean,
+    agentId?: string,
+    samplingRate: number,
+    callEvent: string,
+    activeSpeaker: string,
 };
 
 export type SocketCallData = {
