+++ lma-websocket-stack/source/app/src/index.ts
@@ -16,16 +16,23 @@ import BlockStream from 'block-stream2';
 import fs from 'fs';
 import { randomUUID } from 'crypto';
 
-import {  
-    startTranscribe, 
-    CallMetaData, 
+import {
+    startTranscribe,
     writeCallStartEvent,
     writeCallEndEvent,
-    writeCallEvent,
-} from './lca';
-import { CallRecordingEvent, SocketCallData } from './entities-lca';
+    CallMetaData,
+    SocketCallData,
+    writeCallRecordingEvent,
+} from './calleventdata';
 
-import { jwtVerifier } from './jwt-verifier';
+import {
+    createWavHeader,
+    posixifyFilename,
+    normalizeErrorForLogging,
+    getClientIP,
+} from './utils';
+
+import { jwtVerifier } from './utils/jwt-verifier';
 
 const AWS_REGION = process.env['AWS_REGION'] || 'us-east-1';
 const RECORDINGS_BUCKET_NAME = process.env['RECORDINGS_BUCKET_NAME'] || undefined;
@@ -33,10 +40,10 @@ const RECORDING_FILE_PREFIX = process.env['RECORDING_FILE_PREFIX'] || 'lca-audio
 const CPU_HEALTH_THRESHOLD = parseInt(process.env['CPU_HEALTH_THRESHOLD'] || '50', 10);
 const LOCAL_TEMP_DIR = process.env['LOCAL_TEMP_DIR'] || '/tmp/';
 const WS_LOG_LEVEL = process.env['WS_LOG_LEVEL'] || 'debug';
+const WS_LOG_INTERVAL = parseInt(process.env['WS_LOG_INTERVAL'] || '120', 10);
+const SHOULD_RECORD_CALL = (process.env['SHOULD_RECORD_CALL'] || 'false') === 'true';
 
 const s3Client = new S3Client({ region: AWS_REGION });
-
-// const isDev = process.env['NODE_ENV'] !== 'PROD';
 
 const socketMap = new Map<WebSocket, SocketCallData>();
 
@@ -49,8 +56,9 @@ const server = fastify({
             translateTime: 'SYS:HH:MM:ss.l',
             colorize: false,
             levelFirst: true,
-        }
+        },
     },
+    disableRequestLogging: true,
 });
 // register the @fastify/websocket plugin with the fastify server
 server.register(websocket);
@@ -58,8 +66,8 @@ server.register(websocket);
 // Setup preHandler hook to authenticate 
 server.addHook('preHandler', async (request, reply) => {
     if (!request.url.includes('health')) { 
-        server.log.debug('Received preHandler hook for authentication. Calling jwtVerifier to authenticate.');
-        server.log.debug(`Websocket Request - URI: <${request.url}>, SocketRemoteAddr: ${request.socket.remoteAddress}, Headers: ${JSON.stringify(request.headers)}`);
+        const clientIP = getClientIP(request.headers);
+        server.log.debug(`[AUTH]: [${clientIP}] - Received preHandler hook for authentication. URI: <${request.url}>, Headers: ${JSON.stringify(request.headers)}`);
     
         await jwtVerifier(request, reply);
     }
@@ -67,93 +75,82 @@ server.addHook('preHandler', async (request, reply) => {
 
 // Setup Route for websocket connection
 server.get('/api/v1/ws', { websocket: true, logLevel: 'debug' }, (connection, request) => {
-    server.log.debug('Received Connection request.');
-    server.log.debug(`Websocket Request - URI: <${request.url}>, SocketRemoteAddr: ${request.socket.remoteAddress}, Headers: ${JSON.stringify(request.headers)}`);
+    const clientIP = getClientIP(request.headers);
+    server.log.debug(`[NEW CONNECTION]: [${clientIP}] - Received new connection request @ /api/v1/ws. URI: <${request.url}>, Headers: ${JSON.stringify(request.headers)}`);
 
-    registerHandlers(connection.socket); // setup the handler functions for websocket events
+    registerHandlers(clientIP, connection.socket); // setup the handler functions for websocket events
 });
 
+type HealthCheckRemoteInfo = {
+    addr: string;
+    tsFirst: number;
+    tsLast: number;
+    count: number;
+};
+const healthCheckStats = new Map<string, HealthCheckRemoteInfo>();
+
 // Setup Route for health check 
-server.get('/health/check', { logLevel: 'debug' }, (request, response) => {
-    // server.log.debug('Received Health Check request.');
-    // server.log.debug(`Websocket Request - URI: <${request.url}>, SocketRemoteAddr: ${request.socket.remoteAddress}, Headers: ${JSON.stringify(request.headers)}`);
+server.get('/health/check', { logLevel: 'warn' }, (request, response) => {
 
+    const now = Date.now();
     const cpuUsage = os.loadavg()[0] / os.cpus().length * 100;
-
     const isHealthy = cpuUsage > CPU_HEALTH_THRESHOLD ? false : true;
     const status = isHealthy ? 200 : 503;
+    
+    const remoteIp = request.socket.remoteAddress || 'unknown';
+    const item = healthCheckStats.get(remoteIp);
+    if (!item) {
+        server.log.debug(`[HEALTH CHECK]: [${remoteIp}] - Received First health check from load balancer. URI: <${request.url}>, Headers: ${JSON.stringify(request.headers)} ==> Health Check status - CPU Usage%: ${cpuUsage}, IsHealthy: ${isHealthy}, Status: ${status}`);
+        healthCheckStats.set(remoteIp, { addr: remoteIp, tsFirst: now, tsLast: now, count: 1 });
+    } else {
+        item.tsLast = now;
+        ++item.count;
+        const elapsed_seconds = Math.round((item.tsLast - item.tsFirst) / 1000);
+        if ((elapsed_seconds % WS_LOG_INTERVAL) == 0) {
+            server.log.debug(`[HEALTH CHECK]: [${remoteIp}] - Received Health check # ${item.count} from load balancer. URI: <${request.url}>, Headers: ${JSON.stringify(request.headers)} ==> Health Check status - CPU Usage%: ${cpuUsage}, IsHealthy: ${isHealthy}, Status: ${status}`);
+        }
+    }
 
-    server.log.debug(`Health Check request received. CPU Usage%: ${cpuUsage}, IsHealthy: ${isHealthy}, Status: ${status}`);
     response
         .code(status)
         .header('Cache-Control', 'max-age=0, no-cache, no-store, must-revalidate, proxy-revalidate')
         .send({ 'Http-Status': status, 'Healthy': isHealthy });
 });
 
-// Start the websocket server on default port 3000 if no port supplied in environment variables
-server.listen(
-    { 
-        port: parseInt(process.env?.['SERVERPORT'] ?? '8080'),
-        host: process.env?.['SERVERHOST'] ?? '127.0.0.1'
-    },
-    (err) => {
-        if (err) {
-            server.log.error('Error starting websocket server: ',err);
-            process.exit(1);
-        }
-        server.log.debug('Websocket server is ready and listening.');
-        server.log.debug(`Routes: \n${server.printRoutes()}`);
-    }
-);
 
 // Setup handlers for websocket events - 'message', 'close', 'error'
-const registerHandlers = (ws: WebSocket): void => {
+const registerHandlers = (clientIP: string, ws: WebSocket): void => {
     ws.on('message', async (data, isBinary): Promise<void> => {
         try {
             if (isBinary) {
                 const audioinput = Buffer.from(data as Uint8Array);
-                await onBinaryMessage(ws, audioinput);
+                await onBinaryMessage(clientIP, ws, audioinput);
             } else {
-                await onTextMessage(ws, Buffer.from(data as Uint8Array).toString('utf8'));
+                await onTextMessage(clientIP, ws, Buffer.from(data as Uint8Array).toString('utf8'));
             }
-        } catch (err) {
-            server.log.error(`Error processing message: ${err}`);
+        } catch (error) {
+            server.log.error(`[ON MESSAGE]: [${clientIP}] - Error processing message: ${normalizeErrorForLogging(error)}`);
             process.exit(1);
         }
     });
 
     ws.on('close', (code: number) => {
-        server.log.debug('Received Websocket close message from the client. Closing the connection.');
+        server.log.debug(`[ON WSCLOSE]: [${clientIP}] Received Websocket close message from the client. Closing the connection.`);
+
         try {
             onWsClose(ws, code);
         } catch (err) {
-            server.log.error('Error in WS close handler: ', err);
+            server.log.error(`[ON WSCLOSE]: [${clientIP}] Error in WS close handler: ${normalizeErrorForLogging(err)}`);
         }
     });
 
     ws.on('error', (error: Error) => {
-        server.log.error('Websocket error, forcing close: ', error);
+        server.log.error(`[ON WSERROR]: [${clientIP}] - Websocket error, forcing close: ${normalizeErrorForLogging(error)}`);
         ws.close();
     });
 };
 
-const posixifyFilename = function (filename: string) {
-    // Replace all invalid characters with underscores.
-    const regex = /[^a-zA-Z0-9_.]/g;
-    const posixFilename = filename.replace(regex, '_');
-    // Remove leading and trailing underscores.
-    return posixFilename.replace(/^_+/g, '').replace(/_+$/g, '');
-};
-
-const getTempRecordingFileName = (callMetaData: CallMetaData): string => {
-    return `${posixifyFilename(callMetaData.callId)}.raw`;
-};
-
-const getWavRecordingFileName = (callMetaData: CallMetaData): string => {
-    return `${posixifyFilename(callMetaData.callId)}.wav`;
-};
-
-const onBinaryMessage = async (ws: WebSocket, data: Uint8Array): Promise<void> => {
+const onBinaryMessage = async (clientIP: string, ws: WebSocket, data: Uint8Array): Promise<void> => {
     // server.log.debug(`Received Binary Message of length ${data.byteLength}`);
 
     const socketData = socketMap.get(ws);
@@ -164,18 +161,19 @@ const onBinaryMessage = async (ws: WebSocket, data: Uint8Array): Promise<void> =
         socketData.writeRecordingStream.write(data);
         socketData.recordingFileSize += data.length;
     } else {
-        server.log.error('Error: received audio data before metadata');
+        server.log.error(`[ON BINARY MESSAGE]: [${clientIP}] - Error: received audio data before metadata. Check logs for errors in START event.`);
     }
 };
 
-const onTextMessage = async (ws: WebSocket, data: string): Promise<void> => {
+const onTextMessage = async (clientIP: string, ws: WebSocket, data: string): Promise<void> => {
     server.log.debug(`Received Text Message of length ${data.length}. Message = ${data}`);
     
-    const callMetaData:CallMetaData = JSON.parse(data);
+    const callMetaData: CallMetaData = JSON.parse(data);
+    
     try {
-        server.log.info(`Call Metadata received from client :  ${data}`);
+        server.log.debug(`[ON TEXT MESSAGE]: [${clientIP}][${callMetaData.callId}] - Call Metadata received from client: ${data}`);
     } catch (error) {
-        server.log.error(`Error parsing call metadata: ${error} ${data}`);
+        server.log.error(`[ON TEXT MESSAGE]: [${clientIP}][${callMetaData.callId}] - Error parsing call metadata: ${data} ${normalizeErrorForLogging(error)}`);
         callMetaData.callId = randomUUID();
     }
     
@@ -184,10 +182,14 @@ const onTextMessage = async (ws: WebSocket, data: string): Promise<void> => {
         callMetaData.callId = callMetaData.callId || randomUUID();
         callMetaData.fromNumber = callMetaData.fromNumber || 'Customer Phone';
         callMetaData.toNumber = callMetaData.toNumber || 'System Phone';
-        callMetaData.shouldRecordCall = callMetaData.shouldRecordCall || false;
+        
+        if (typeof callMetaData.shouldRecordCall === 'undefined' || callMetaData.shouldRecordCall === null) {
+            callMetaData.shouldRecordCall = SHOULD_RECORD_CALL;
+        }  
+
         callMetaData.agentId = callMetaData.agentId || randomUUID();  
 
-        await writeCallStartEvent(callMetaData);
+        await writeCallStartEvent(callMetaData, server);
         const tempRecordingFilename = getTempRecordingFileName(callMetaData);
         // wavFileName = `${callMetaData.callId}.wav`;
         const writeRecordingStream = fs.createWriteStream(path.join(LOCAL_TEMP_DIR, tempRecordingFilename));
@@ -206,24 +208,26 @@ const onTextMessage = async (ws: WebSocket, data: string): Promise<void> => {
             ended: false
         };
         socketMap.set(ws, socketCallMap);
-        startTranscribe(callMetaData, audioInputStream, socketCallMap);
+        startTranscribe(callMetaData, audioInputStream, socketCallMap, server);
 
     } else if (callMetaData.callEvent === 'SPEAKER_CHANGE') {
         const socketData = socketMap.get(ws);
+        server.log.debug(`[${callMetaData.callEvent}]: [${callMetaData.callId}] - Received speaker change. Active speaker = ${callMetaData.activeSpeaker}`);
+
         if (socketData && socketData.callMetadata) {
             socketData.callMetadata.activeSpeaker = callMetaData.activeSpeaker;
         } else {
             // this is not a valid call
-            server.log.error(`Invalid call : ${JSON.stringify(callMetaData)}`);
+            server.log.error(`[${callMetaData.callEvent}]: [${callMetaData.callId}] - Invalid call : ${JSON.stringify(callMetaData)}`);
         }
     } else if (callMetaData.callEvent === 'END') {
         const socketData = socketMap.get(ws);
         if (!socketData || !(socketData.callMetadata)) {
-            server.log.debug(`Received END without having a call:  ${JSON.stringify(callMetaData)}`);
+            server.log.error(`[${callMetaData.callEvent}]: [${callMetaData.callId}] - Received END without starting a call:  ${JSON.stringify(callMetaData)}`);
             return;
         }
-        server.log.debug(`Received call end event from client, writing it to KDS:  ${JSON.stringify(callMetaData)}`);
-        await endCall(ws, callMetaData, socketData);
+        server.log.debug(`[${callMetaData.callEvent}]: [${callMetaData.callId}] - Received call end event from client, writing it to KDS:  ${JSON.stringify(callMetaData)}`);
+        await endCall(ws, socketData, callMetaData);
     }
 };
 
@@ -232,12 +236,12 @@ const onWsClose = async (ws: WebSocket, code: number): Promise<void> => {
     ws.close(code);
     const socketData = socketMap.get(ws);
     if (socketData) {
-        server.log.debug(`Writing call end event due to websocket close event ${JSON.stringify(socketData.callMetadata)}`);
-        await endCall(ws, undefined, socketData);
+        server.log.debug(`[ON WSCLOSE]: [${socketData.callMetadata.callId}] - Writing call end event due to websocket close event ${JSON.stringify(socketData.callMetadata)}`);
+        await endCall(ws, socketData);
     }
 };
 
-const endCall = async (ws: WebSocket, callMetaData: CallMetaData|undefined, socketData: SocketCallData): Promise<void> => {
+const endCall = async (ws: WebSocket, socketData: SocketCallData, callMetaData?: CallMetaData): Promise<void> => {
     
     if (callMetaData === undefined) {
         callMetaData = socketData.callMetadata;
@@ -246,54 +250,60 @@ const endCall = async (ws: WebSocket, callMetaData: CallMetaData|undefined, sock
     if (socketData !== undefined && socketData.ended === false) {
         socketData.ended = true;
 
-        await writeCallEndEvent(callMetaData);
-        if (socketData.writeRecordingStream && socketData.recordingFileSize) {
-            socketData.writeRecordingStream.end();
-            const header = createHeader(callMetaData, socketData.recordingFileSize);
-            const tempRecordingFilename = getTempRecordingFileName(callMetaData);
-            const wavRecordingFilename = getWavRecordingFileName(callMetaData);
-            const readStream = fs.createReadStream(path.join(LOCAL_TEMP_DIR, tempRecordingFilename));
-            const writeStream = fs.createWriteStream(path.join(LOCAL_TEMP_DIR, wavRecordingFilename));
-            writeStream.write(header);
-            for await (const chunk of readStream) {
-                writeStream.write(chunk);
-            }
-            writeStream.end();
-    
-            await writeToS3(tempRecordingFilename);
-            await writeToS3(wavRecordingFilename);
-            await deleteTempFile(path.join(LOCAL_TEMP_DIR, tempRecordingFilename));
-            await deleteTempFile(path.join(LOCAL_TEMP_DIR, wavRecordingFilename));
-    
-            const url = new URL(RECORDING_FILE_PREFIX + wavRecordingFilename, `https://${RECORDINGS_BUCKET_NAME}.s3.${AWS_REGION}.amazonaws.com`);
-            const recordingUrl = url.href;
-            
-            const callEvent: CallRecordingEvent = {
-                EventType: 'ADD_S3_RECORDING_URL',
-                CallId: callMetaData.callId,
-                RecordingUrl: recordingUrl
-            };
-            await writeCallEvent(callEvent);
-        }
-        if (socketData.audioInputStream) {
-            server.log.info(`Closing audio input stream:  ${JSON.stringify(callMetaData)}`);
-            socketData.audioInputStream.end();
-            socketData.audioInputStream.destroy();
-        }
-        if (socketData) {
-            server.log.info(`Deleting websocket from map: ${JSON.stringify(callMetaData)}`);
-            socketMap.delete(ws);
-        }
-    } else {
-        server.log.info(`Already received the end call event: ${JSON.stringify(callMetaData)}`);
+        if (typeof callMetaData !== undefined && callMetaData != null) {
+            await writeCallEndEvent(callMetaData, server);
+            if (socketData.writeRecordingStream && socketData.recordingFileSize) {
+                socketData.writeRecordingStream.end();
 
+                if (callMetaData.shouldRecordCall) {
+                    const header = createWavHeader(callMetaData.samplingRate, socketData.recordingFileSize);
+                    const tempRecordingFilename = getTempRecordingFileName(callMetaData);
+                    const wavRecordingFilename = getWavRecordingFileName(callMetaData);
+                    const readStream = fs.createReadStream(path.join(LOCAL_TEMP_DIR, tempRecordingFilename));
+                    const writeStream = fs.createWriteStream(path.join(LOCAL_TEMP_DIR, wavRecordingFilename));
+                    writeStream.write(header);
+                    for await (const chunk of readStream) {
+                        writeStream.write(chunk);
+                    }
+                    writeStream.end();
+        
+                    await writeToS3(callMetaData, tempRecordingFilename);
+                    await writeToS3(callMetaData, wavRecordingFilename);
+                    await deleteTempFile(callMetaData, path.join(LOCAL_TEMP_DIR, tempRecordingFilename));
+                    await deleteTempFile(callMetaData, path.join(LOCAL_TEMP_DIR, wavRecordingFilename));
+        
+                    const url = new URL(RECORDING_FILE_PREFIX + wavRecordingFilename, `https://${RECORDINGS_BUCKET_NAME}.s3.${AWS_REGION}.amazonaws.com`);
+                    const recordingUrl = url.href;
+                
+                    await writeCallRecordingEvent(callMetaData, recordingUrl, server);
+                }
+            }
+        
+            if (socketData.audioInputStream) {
+                server.log.debug(`[${callMetaData.callEvent}]: [${callMetaData.callId}] - Closing audio input stream:  ${JSON.stringify(callMetaData)}`);
+                socketData.audioInputStream.end();
+                socketData.audioInputStream.destroy();
+            }
+            if (socketData) {
+                server.log.debug(`[${callMetaData.callEvent}]: [${callMetaData.callId}] - Deleting websocket from map: ${JSON.stringify(callMetaData)}`);
+                socketMap.delete(ws);
+            }
+        } else {
+            server.log.error('[END]: Missing Call Meta Data in END event');
+        }
+
+    } else {
+        if (typeof callMetaData !== undefined && callMetaData != null) {
+            server.log.error(`[${callMetaData.callEvent}]: [${callMetaData.callId}] - Duplicate End call event. Already received the end call event: ${JSON.stringify(callMetaData)}`);
+        } else {
+            server.log.error('[END]: Duplicate End call event. Missing Call Meta Data in END event');
+
+        }
     }
 };
 
-const writeToS3 = async (tempFileName:string) => {
+const writeToS3 = async (callMetaData: CallMetaData, tempFileName: string) => {
     const sourceFile = path.join(LOCAL_TEMP_DIR, tempFileName);
-
-    server.log.debug(`Uploading audio to S3: ${sourceFile}`);
 
     let data;
     const fileStream = fs.createReadStream(sourceFile);
@@ -304,9 +314,9 @@ const writeToS3 = async (tempFileName:string) => {
     };
     try {
         data = await s3Client.send(new PutObjectCommand(uploadParams));
-        server.log.debug(`Uploading to S3 complete: ${JSON.stringify(data)}`);
+        server.log.debug(`[${callMetaData.callEvent}]: [${callMetaData.callId}] - Uploaded ${sourceFile} to S3 complete: ${JSON.stringify(data)}`);
     } catch (err) {
-        server.log.error(`S3 upload error: ${JSON.stringify(err)}`);
+        server.log.error(`[${callMetaData.callEvent}]: [${callMetaData.callId}] - Error uploading ${sourceFile} to S3: ${normalizeErrorForLogging(err)}`);
 
     } finally {
         fileStream.destroy();
@@ -314,42 +324,35 @@ const writeToS3 = async (tempFileName:string) => {
     return data;
 };
 
-const deleteTempFile = async(sourceFile:string) => {
+const getTempRecordingFileName = (callMetaData: CallMetaData): string => {
+    return `${posixifyFilename(callMetaData.callId)}.raw`;
+};
+
+const getWavRecordingFileName = (callMetaData: CallMetaData): string => {
+    return `${posixifyFilename(callMetaData.callId)}.wav`;
+};
+
+const deleteTempFile = async(callMetaData:CallMetaData, sourceFile:string) => {
     try {
         await fs.promises.unlink(sourceFile);
+        server.log.debug(`[${callMetaData.callEvent}]: [${callMetaData.callId}] - Deleted tmp file ${sourceFile}`);
     } catch (err) {
-        server.log.error(`Error deleting temp file:  ${JSON.stringify(err)}`);
+        server.log.error(`[${callMetaData.callEvent}]: [${callMetaData.callId}] - Error deleting tmp file ${sourceFile} : ${normalizeErrorForLogging(err)}`);
     }
 };
 
-const createHeader = function createHeader(callMetaData:CallMetaData, length:number) {
-    const buffer = Buffer.alloc(44);
-  
-    // RIFF identifier 'RIFF'
-    buffer.writeUInt32BE(1380533830, 0);
-    // file length minus RIFF identifier length and file description length
-    buffer.writeUInt32LE(36 + length, 4);
-    // RIFF type 'WAVE'
-    buffer.writeUInt32BE(1463899717, 8);
-    // format chunk identifier 'fmt '
-    buffer.writeUInt32BE(1718449184, 12);
-    // format chunk length
-    buffer.writeUInt32LE(16, 16);
-    // sample format (raw)
-    buffer.writeUInt16LE(1, 20);
-    // channel count
-    buffer.writeUInt16LE(2, 22);
-    // sample rate
-    buffer.writeUInt32LE(callMetaData.samplingRate, 24);
-    // byte rate (sample rate * block align)
-    buffer.writeUInt32LE(callMetaData.samplingRate * 2 * 2, 28);
-    // block align (channel count * bytes per sample)
-    buffer.writeUInt16LE(2 * 2, 32);
-    // bits per sample
-    buffer.writeUInt16LE(16, 34);
-    // data chunk identifier 'data'
-    buffer.writeUInt32BE(1684108385, 36);
-    buffer.writeUInt32LE(length, 40);
-  
-    return buffer;
-};
+// Start the websocket server on default port 3000 if no port supplied in environment variables
+server.listen(
+    { 
+        port: parseInt(process.env?.['SERVERPORT'] ?? '8080'),
+        host: process.env?.['SERVERHOST'] ?? '127.0.0.1'
+    },
+    (err) => {
+        if (err) {
+            server.log.error(`[WS SERVER STARTUP]: Error starting websocket server: ${normalizeErrorForLogging(err)}`);
+            process.exit(1);
+        }
+        server.log.debug('[WS SERVER STARTUP]: Websocket server is ready and listening.');
+        server.log.info(`[[WS SERVER STARTUP]]: Routes: \n${server.printRoutes()}`);
+    }
+);
