+++ lma-websocket-stack/source/app/src/utils/wav.ts
@@ -1,0 +1,91 @@
+
+export const createWavHeader = (sampleRate: number, length: number) => {
+    const buffer = Buffer.alloc(44);
+  
+    // RIFF identifier 'RIFF'
+    buffer.writeUInt32BE(1380533830, 0);
+    // file length minus RIFF identifier length and file description length
+    buffer.writeUInt32LE(36 + length, 4);
+    // RIFF type 'WAVE'
+    buffer.writeUInt32BE(1463899717, 8);
+    // format chunk identifier 'fmt '
+    buffer.writeUInt32BE(1718449184, 12);
+    // format chunk length
+    buffer.writeUInt32LE(16, 16);
+    // sample format (raw)
+    buffer.writeUInt16LE(1, 20);
+    // channel count
+    buffer.writeUInt16LE(2, 22);
+    // sample rate
+    buffer.writeUInt32LE(sampleRate, 24);
+    // byte rate (sample rate * block align)
+    buffer.writeUInt32LE(sampleRate * 2 * 2, 28);
+    // block align (channel count * bytes per sample)
+    buffer.writeUInt16LE(2 * 2, 32);
+    // bits per sample
+    buffer.writeUInt16LE(16, 34);
+    // data chunk identifier 'data'
+    buffer.writeUInt32BE(1684108385, 36);
+    buffer.writeUInt32LE(length, 40);
+  
+    return buffer;
+};
+
+
+/**
+ * Lookup table to convert u-Law bytes to their Linear-16 sample values
+ */
+const ulawToL16Lut = new Int16Array([
+    -32124, -31100, -30076, -29052, -28028, -27004, -25980, -24956,
+    -23932, -22908, -21884, -20860, -19836, -18812, -17788, -16764,
+    -15996, -15484, -14972, -14460, -13948, -13436, -12924, -12412,
+    -11900, -11388, -10876, -10364, -9852, -9340, -8828, -8316,
+    -7932, -7676, -7420, -7164, -6908, -6652, -6396, -6140,
+    -5884, -5628, -5372, -5116, -4860, -4604, -4348, -4092,
+    -3900, -3772, -3644, -3516, -3388, -3260, -3132, -3004,
+    -2876, -2748, -2620, -2492, -2364, -2236, -2108, -1980,
+    -1884, -1820, -1756, -1692, -1628, -1564, -1500, -1436,
+    -1372, -1308, -1244, -1180, -1116, -1052, -988, -924,
+    -876, -844, -812, -780, -748, -716, -684, -652,
+    -620, -588, -556, -524, -492, -460, -428, -396,
+    -372, -356, -340, -324, -308, -292, -276, -260,
+    -244, -228, -212, -196, -180, -164, -148, -132,
+    -120, -112, -104, -96, -88, -80, -72, -64,
+    -56, -48, -40, -32, -24, -16, -8, -1,
+    32124, 31100, 30076, 29052, 28028, 27004, 25980, 24956,
+    23932, 22908, 21884, 20860, 19836, 18812, 17788, 16764,
+    15996, 15484, 14972, 14460, 13948, 13436, 12924, 12412,
+    11900, 11388, 10876, 10364, 9852, 9340, 8828, 8316,
+    7932, 7676, 7420, 7164, 6908, 6652, 6396, 6140,
+    5884, 5628, 5372, 5116, 4860, 4604, 4348, 4092,
+    3900, 3772, 3644, 3516, 3388, 3260, 3132, 3004,
+    2876, 2748, 2620, 2492, 2364, 2236, 2108, 1980,
+    1884, 1820, 1756, 1692, 1628, 1564, 1500, 1436,
+    1372, 1308, 1244, 1180, 1116, 1052, 988, 924,
+    876, 844, 812, 780, 748, 716, 684, 652,
+    620, 588, 556, 524, 492, 460, 428, 396,
+    372, 356, 340, 324, 308, 292, 276, 260,
+    244, 228, 212, 196, 180, 164, 148, 132,
+    120, 112, 104, 96, 88, 80, 72, 64,
+    56, 48, 40, 32, 24, 16, 8, 0
+]);
+
+
+/**
+ * Decodes an array of audio samples encoded with u-Law to Linear16
+ * 
+ * @param {Uint8Array} ulawBuf Array of u-Law bytes to convert to Linear16
+ * @returns {Int16Array} Array of Linear16 samples [-32768, 32767]
+ */
+export const ulawToL16 = (ulawBuf: Uint8Array): Int16Array => {
+    const size = ulawBuf.length;
+    const res = new Int16Array(size);
+    for (let i = 0; i < size; ++i) {
+        res[i] = ulawToL16Lut[ulawBuf[i]];
+    }
+    return res;
+};
+
+export const msToBytes = (ms: number, samplerate: number, samplebytes: number): number => {
+    return samplerate * (ms / 1000) * samplebytes;
+};
